who_scored/requirements.txt
beautifulsoup4==4.12.2
selenium==4.10.0
webdriver-manager==3.8.6
pyyaml
pydantic==1.10.7
pre-commit==3.3.1
pytest==7.3.1
black==22.3.0
flake8==6.0.0
pylint==2.17.4